Adds session controller
This change create the session controller and its associated logic.  It
also finalizes the User model and creates a form object to handle the
login form.

I created a view based on the mockup for the login page and I updated
the main template view to handle bootstrap.

I stubbed out the routes for the application and commented out the ones
that aren't used yet.