Adds database connection
This setups the database backend for the project using a PosgreSQL
database hosted on Heroku.  I was tempted to do the whole thing using
Sqlite or H2 for a file that could be stored in the repo but decided
that using one of those isn't representative of an actual deployment.

In addition to getting the DB connection setup, this commit also enables
the EBean ORM and evolutions.  Evolutions are database migrations
written in SQL and automatically applied.