Adds employeetype to User
I don't know how I forgot this but each employee needs to be assigned as
an employee type for authorizing the various pages.  An employee should
only be able to submit requests.  A supervisor should be able to
approve/deny requests.  An HR person should be able to modify users.