Adds authorization to project
This change adds methods in the Application controller that check if the
current user is authorized to view the page and either run the lambda
passed for success or redirect and show an error message.  I'm pretty
proud that I saw a place to use a higher-order function and used it.

It also updates the other controllers to use these methods for
authorizing users to the various parts of the website.