Adds a bar for the flash and shows user info
This change does two minor things both related to the view.  First, it
adds a new template for the flash.  It either shows a warning or
success depending on what's in the flash.  I guess, it's technically
possible to show both but I don't see that as a real possibility.

The second thing this does is adds a new parameter to the vacation
request index, main, and navbar templates for the current user.  This
is so on the vacation request index page, I can show the user's name
and type in the navbar.