Adds supervisor index page
Instead of jamming a lot of logic about who to show what in the request
index view, I've moved it to the controller and created a new view just
for supervisors.  They can see the list of requests and when they click
on one, it takes them straight to the edit page.